# dashboard/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmOut": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "count",
          "bytes",
          "token"
        ],
        "title": "ConfirmOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetentionBody": {
        "properties": {
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          },
          "max_age_days": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Age Days"
          },
          "max_vault_gb": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Vault Gb"
          }
        },
        "title": "RetentionBody",
        "type": "object"
      },
      "RetentionOut": {
        "properties": {
          "env_pinned": {
            "title": "Env Pinned",
            "type": "boolean"
          },
          "max_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Age Days"
          },
          "max_vault_gb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Vault Gb"
          },
          "vault_used_bytes": {
            "title": "Vault Used Bytes",
            "type": "integer"
          }
        },
        "required": [
          "max_age_days",
          "max_vault_gb",
          "env_pinned",
          "vault_used_bytes"
        ],
        "title": "RetentionOut",
        "type": "object"
      },
      "RetentionSaveOut": {
        "properties": {
          "confirm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "retention": {
            "$ref": "#/components/schemas/RetentionOut"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "saved",
          "confirm",
          "retention"
        ],
        "title": "RetentionSaveOut",
        "type": "object"
      },
...
      "ScheduleBody": {
        "properties": {
          "archive_enabled": {
            "title": "Archive Enabled",
            "type": "boolean"
          },
          "archive_overlap_hours": {
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Archive Overlap Hours",
            "type": "integer"
          },
          "archive_time": {
            "title": "Archive Time",
            "type": "string"
          },
          "backfill_days": {
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Backfill Days",
            "type": "integer"
          },
          "backfill_dow": {
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Backfill Dow",
            "type": "integer"
          },
          "backfill_enabled": {
            "title": "Backfill Enabled",
            "type": "boolean"
          },
          "backfill_time": {
            "title": "Backfill Time",
            "type": "string"
          },
          "integrity_scan_enabled": {
            "title": "Integrity Scan Enabled",
            "type": "boolean"
          },
          "integrity_scan_sample_pct": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Integrity Scan Sample Pct",
            "type": "number"
          },
          "reconcile_enabled": {
            "title": "Reconcile Enabled",
            "type": "boolean"
          },
          "reconcile_interval_hours": {
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Reconcile Interval Hours",
            "type": "integer"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "required": [
          "archive_enabled",
          "archive_time",
          "archive_overlap_hours",
          "backfill_enabled",
          "backfill_dow",
          "backfill_time",
          "backfill_days",
          "reconcile_enabled",
          "reconcile_interval_hours",
          "integrity_scan_enabled",
          "integrity_scan_sample_pct"
        ],
        "title": "ScheduleBody",
        "type": "object"
      },
      "ScheduleOut": {
        "properties": {
          "archive_enabled": {
            "title": "Archive Enabled",
            "type": "boolean"
          },
          "archive_overlap_hours": {
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Archive Overlap Hours",
            "type": "integer"
          },
          "archive_time": {
            "title": "Archive Time",
            "type": "string"
          },
          "backfill_days": {
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Backfill Days",
            "type": "integer"
          },
          "backfill_dow": {
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Backfill Dow",
            "type": "integer"
          },
          "backfill_enabled": {
            "title": "Backfill Enabled",
            "type": "boolean"
          },
          "backfill_time": {
            "title": "Backfill Time",
            "type": "string"
          },
          "env_pinned": {
            "title": "Env Pinned",
            "type": "boolean"
          },
          "integrity_scan_enabled": {
            "title": "Integrity Scan Enabled",
            "type": "boolean"
          },
          "integrity_scan_sample_pct": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Integrity Scan Sample Pct",
            "type": "number"
          },
          "reconcile_enabled": {
            "title": "Reconcile Enabled",
            "type": "boolean"
          },
          "reconcile_interval_hours": {
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Reconcile Interval Hours",
            "type": "integer"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "required": [
          "archive_enabled",
          "archive_time",
          "archive_overlap_hours",
          "backfill_enabled",
          "backfill_dow",
          "backfill_time",
          "backfill_days",
          "reconcile_enabled",
          "reconcile_interval_hours",
          "integrity_scan_enabled",
          "integrity_scan_sample_pct",
          "env_pinned"
        ],
        "title": "ScheduleOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ReoVault",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/retention": {
      "get": {
        "operationId": "get_retention_api_v1_retention_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Retention",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_retention_api_v1_retention_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSaveOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Retention",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/schedule": {
      "get": {
        "operationId": "get_schedule_api_v1_schedule_get",
        "parameters": [
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schedule",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_schedule_api_v1_schedule_put",
        "parameters": [
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Schedule",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}